fix(ci): restore main checks
